feat: optional auto-open disclosure triangle for PR comment detail views (#449)
* feat: optional auto-open disclosure triangle for PR comment detail views

This option will allow teams who prefer it, to draw more attention
to the short resource diff and/or the detailed log views in the
PR comment, to highlight potentially destructive changes more readily.

The default behavior of these new optional inputs matches the prior
behavior, so there is no change for teams who do not opt in.

* fix: quote syntax for `comment-pr-open-` inputs

* rename, document and test input variables

* concise description

* consistent spacing

---------

Co-authored-by: Rishav Dhar <19497993+rdhar@users.noreply.github.com>